+++ src/soyokaze/mainwindow/state/MainWindowSearchingState.h
@@ -13,7 +13,7 @@ public:
 	explicit SearchingState(LauncherWindowStateContextIF* context, bool allowParamSearching = true);
 
 	/** ウインドウの非表示要求を受けたとき、非表示Stateへ遷移する */
-	void OnDeactivate() override;
+	void OnHideRequested() override;
 	/** 現在の候補の実行要求を受けたとき、実行し、ウインドウが閉じた場合は非表示Stateへ遷移する */
 	void OnExecuteRequested() override;
 	/** キャンセル操作を受けたとき、入力内容をクリアして待機中Stateへ戻る */